Catch integrity error and show form validation error
Should only happen if an account was created with the proposed email
address after the email change request was made. In this edge case we
can assume that the authenticated user is the one who wants to change
the email address of his/her account, because the confirm link has
been clicked as well.
They might have accidentally created a new account with the proposed
email address in the meantime.